anwsers for lesson 10

+++ controllers/comments.js
@@ -12,7 +12,7 @@ const getComments = async (req, res) => {
     throw new NotFoundError('This post does not exist.');
   }
 
-  const commentsData = await Comment.find({post: postId}).populate('madeBy'); 
+  const commentsData = await Comment.find({post: postId}).populate('madeBy').sort('madeBy'); 
   
   if (commentsData.length === 0) {
     return res.status(StatusCodes.OK).json({message: 'Be the first to comment'});
@@ -30,13 +30,16 @@ const getComments = async (req, res) => {
 };
 
 const getComment = async (req, res) => {
-  const post = await Post.find({_id: req.body.post});
+  const {
+    params: {postId, commentId} 
+  } = req;
+
+  const post = await Post.find({_id: postId});
 
   if (post.length === 0) {
     throw new NotFoundError('This post does not exist.');
   }
-  
-  const commentId = req.params.commentId;
+   
   const comment = await Comment.find({_id: commentId}); 
   
   if (!comment) {
@@ -46,33 +49,40 @@ const getComment = async (req, res) => {
   res.status(StatusCodes.OK).json({comment});
 };
 
-const  createComment = async (req, res) => { 
-  if (!req.body.text) {
+const  createComment = async (req, res) => {  
+  const { 
+    user: {userId},
+    body: {text},
+    params: {postId} 
+  } = req;
+
+  if (!text) {
     throw new BadRequestError('Please provide comment text');
   }
+ 
+  const parentPost = await Post.find({_id: postId});
 
-  req.body.madeBy = req.user.userId; 
-  req.body.post = req.params.postId; 
-
-  const post = await Post.find({_id: req.body.post});
-
-  if (post.length === 0) {
+  if (!parentPost) {
     throw new NotFoundError('This post does not exist.');
-  }
+  } 
 
-  const comment = await Comment.create(req.body); 
+  const comment = await Comment.create({text, madeBy: userId, post: postId}); 
 
   res.status(StatusCodes.CREATED).json({comment});
 }
 
 const editComment = async (req, res) => {
-  if (!req.body.text) {
+  const {
+    user: {userId},
+    body: {text},
+    params: {commentId} 
+  } = req;
+
+  if (!text) {
     throw new BadRequestError('Please provide comment text');
-  }
+  } 
 
-  const commentId = req.params.commentId;
-
-  const comment = await Comment.findOneAndUpdate({_id: commentId}, {text: req.body.text}, {new: true});
+  const comment = await Comment.findOneAndUpdate({_id: commentId, madeBy: userId}, {text}, {new: true, runValidators: true});
 
   if(!comment) {
     throw new NotFoundError('This comment does not exist');
@@ -82,8 +92,12 @@ const editComment = async (req, res) => {
 };
 
 const deleteComment = async (req, res) => {
-  const commentId = req.params.commentId;
-  const comment = await Comment.findOneAndDelete({_id: commentId});
+  const {
+    user: {userId},
+    params: {commentId} 
+  } = req;
+ 
+  const comment = await Comment.findOneAndDelete({_id: commentId, madeBy: userId});
 
   if(!comment) {
     throw new NotFoundError('This comment does not exist');
